feat: sink output plugin into bulider-shared
