+++ packages/builder/builder-shared/src/apply/index.ts
@@ -1,1 +1,2 @@
 export * from './html';
+export * from './output';
